chore: organize utils and remove dependency on path_provider
