+++ packages/noports_core/lib/src/common/validation_utils.dart
@@ -3,76 +3,13 @@ import 'dart:io';
 import 'package:at_chops/at_chops.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_utils/at_utils.dart';
+import 'package:noports_core/src/common/file_system_utils.dart';
 import 'package:path/path.dart' as path;
-
-import 'package:path_provider/path_provider.dart' as path_provider;
-
-/// Get the home directory or null if unknown.
-Future<String> getHomeDirectory() async {
-  String? homeDir;
-  switch (Platform.operatingSystem) {
-    case 'linux':
-    case 'macos':
-      homeDir = Platform.environment['HOME'];
-    case 'windows':
-      homeDir = Platform.environment['USERPROFILE'];
-    case 'android':
-      // android to try external storage first and fallback to the ApplicationSupportDirectory
-      homeDir = await path_provider
-          .getExternalStorageDirectory()
-          .then((dir) => dir?.path);
-    case 'ios':
-    case 'fuchsia':
-    default:
-      // ios and fuchsia to use the ApplicationSupportDirectory
-      homeDir = null;
-  }
-  return homeDir ??
-      await path_provider
-          .getApplicationSupportDirectory()
-          .then((dir) => dir.path);
-}
-
-/// Get the local username or null if unknown
-String? getUserName({bool throwIfNull = false}) {
-  Map<String, String> envVars = Platform.environment;
-  if (!Platform.isWindows) {
-    return envVars['USER'];
-  } else if (Platform.isWindows) {
-    return envVars['USERPROFILE'];
-  }
-  if (throwIfNull) {
-    throw ('\nUnable to determine your username: please set environment variable\n\n');
-  }
-  return null;
-}
-
-Future<bool> fileExists(String file) async {
-  bool f = await File(file).exists();
-  return f;
-}
 
 const String sshnpDeviceNameRegex = r'[a-zA-Z0-9_]{0,15}';
 
 bool checkNonAscii(String test) {
   return RegExp(sshnpDeviceNameRegex).allMatches(test).first.group(0) != test;
-}
-
-String getDefaultAtKeysFilePath(String homeDirectory, String? atSign) {
-  if (atSign == null) return '';
-  return path.normalize('$homeDirectory/.atsign/keys/${atSign}_key.atKeys');
-}
-
-String getDefaultSshDirectory(String homeDirectory) {
-  return path.normalize('$homeDirectory/.ssh/');
-}
-
-String getDefaultSshnpDirectory(String homeDirectory) {
-  return path.normalize('$homeDirectory/.sshnp/');
-}
-
-String getDefaultSshnpConfigDirectory(String homeDirectory) {
-  return path.normalize('$homeDirectory/.sshnp/config');
 }
 
 /// Checks if the provided atSign's atServer has been properly activated with a public RSA key.
@@ -115,110 +52,6 @@ void assertValidValue(Map m, String k, Type t) {
   }
 }
 
-Future<(String, String)> generateSshKeys(
-    {required bool rsa,
-    required String sessionId,
-    String? sshHomeDirectory}) async {
-  sshHomeDirectory ??= getDefaultSshDirectory(await getHomeDirectory());
-  if (!Directory(sshHomeDirectory).existsSync()) {
-    Directory(sshHomeDirectory).createSync();
-  }
-
-  if (rsa) {
-    await Process.run('ssh-keygen',
-        ['-t', 'rsa', '-b', '4096', '-f', '${sessionId}_sshnp', '-q', '-N', ''],
-        workingDirectory: sshHomeDirectory);
-  } else {
-    await Process.run(
-        'ssh-keygen',
-        [
-          '-t',
-          'ed25519',
-          '-a',
-          '100',
-          '-f',
-          '${sessionId}_sshnp',
-          '-q',
-          '-N',
-          ''
-        ],
-        workingDirectory: sshHomeDirectory);
-  }
-
-  String sshPublicKey =
-      await File('$sshHomeDirectory/${sessionId}_sshnp.pub').readAsString();
-  String sshPrivateKey =
-      await File('$sshHomeDirectory/${sessionId}_sshnp').readAsString();
-
-  return (sshPublicKey, sshPrivateKey);
-}
-
-Future<void> addEphemeralKeyToAuthorizedKeys(
-    {required String sshPublicKey,
-    required int localSshdPort,
-    String sessionId = '',
-    String permissions = ''}) async {
-  // Check to see if the ssh public key looks like one!
-  if (!sshPublicKey.startsWith('ssh-')) {
-    throw ('$sshPublicKey does not look like a public key');
-  }
-
-  String homeDirectory = await getHomeDirectory();
-  var sshHomeDirectory = getDefaultSshDirectory(homeDirectory);
-
-  if (!Directory(sshHomeDirectory).existsSync()) {
-    Directory(sshHomeDirectory).createSync();
-  }
-
-  // Check to see if the ssh Publickey is already in the authorized_keys file.
-  // If not, then append it.
-  var authKeys = File(path.normalize('$sshHomeDirectory/authorized_keys'));
-
-  var authKeysContent = await authKeys.readAsString();
-  if (!authKeysContent.endsWith('\n')) {
-    await authKeys.writeAsString('\n', mode: FileMode.append);
-  }
-
-  if (!authKeysContent.contains(sshPublicKey)) {
-    if (permissions.isNotEmpty && !permissions.startsWith(',')) {
-      permissions = ',$permissions';
-    }
-    // Set up a safe authorized_keys file, for the ssh tunnel
-    await authKeys.writeAsString(
-      'command="echo \\"ssh session complete\\";sleep 20"'
-      ',PermitOpen="localhost:$localSshdPort"'
-      '$permissions'
-      ' '
-      '${sshPublicKey.trim()}'
-      ' '
-      'sshnp_ephemeral_$sessionId\n',
-      mode: FileMode.append,
-      flush: true,
-    );
-  }
-}
-
-Future<void> removeEphemeralKeyFromAuthorizedKeys(
-    String sessionId, AtSignLogger logger,
-    {String? sshHomeDirectory}) async {
-  try {
-    sshHomeDirectory ??= getDefaultSshDirectory(await getHomeDirectory());
-    final File file = File(path.normalize('$sshHomeDirectory/authorized_keys'));
-    logger.info('Removing ephemeral key for session $sessionId'
-        ' from ${file.absolute.path}');
-    // read into List of strings
-    final List<String> lines = await file.readAsLines();
-    // find the line we want to remove
-    lines.removeWhere((element) => element.contains(sessionId));
-    // Write back the file and add a \n
-    await file.writeAsString(lines.join('\n'));
-    await file.writeAsString('\n', mode: FileMode.writeOnlyAppend);
-  } catch (e) {
-    logger.severe(
-        'Unable to tidy up ${path.normalize('$sshHomeDirectory/authorized_keys')}');
-  }
-}
-
 String signAndWrapAndJsonEncode(AtClient atClient, Map payload) {
   Map envelope = {'payload': payload};
 
@@ -234,13 +67,14 @@ String signAndWrapAndJsonEncode(AtClient atClient, Map payload) {
 }
 
 Future<void> verifyEnvelopeSignature(AtClient atClient, String requestingAtsign,
-    AtSignLogger logger, Map envelope) async {
+    AtSignLogger logger, Map envelope,
+    {bool useFileStorage = true}) async {
   final String signature = envelope['signature'];
   Map payload = envelope['payload'];
   final hashingAlgo = HashingAlgoType.values.byName(envelope['hashingAlgo']);
   final signingAlgo = SigningAlgoType.values.byName(envelope['signingAlgo']);
   final pk = await getLocallyCachedPK(atClient, requestingAtsign,
-      useFileStorage: true);
+      useFileStorage: useFileStorage);
   AtSigningVerificationInput input = AtSigningVerificationInput(
       jsonEncode(payload), base64Decode(signature), pk)
     ..signingMode = AtSigningMode.data
@@ -292,8 +126,8 @@ Future<String?> _fetchFromLocalPKCache(
     AtClient atClient, String atSign, bool useFileStorage) async {
   String dontAtMe = atSign.substring(1);
   if (useFileStorage) {
-    String fn = path.normalize(
-        '${await getHomeDirectory()}/.atsign/sshnp/cached_pks/$dontAtMe');
+    String fn = path
+        .normalize('${getHomeDirectory()}/.atsign/sshnp/cached_pks/$dontAtMe');
     File f = File(fn);
     if (await f.exists()) {
       return (await f.readAsString()).trim();
@@ -317,7 +151,7 @@ Future<bool> _storeToLocalPKCache(
   String dontAtMe = atSign.substring(1);
   if (useFileStorage) {
     String dirName =
-        path.normalize('${await getHomeDirectory()}/.atsign/sshnp/cached_pks');
+        path.normalize('${getHomeDirectory()}/.atsign/sshnp/cached_pks');
     String fileName = path.normalize('$dirName/$dontAtMe');
 
     File f = File(fileName);
