+++ packages/noports_core/lib/src/sshnp/sshnp_impl/sshnp_local_file_mixin.dart
@@ -1,8 +1,13 @@
+import 'dart:io';
+
 import 'package:noports_core/src/sshnp/sshnp_impl/sshnp_impl.dart';
 import 'package:noports_core/src/sshnp/sshnp_result.dart';
 
 mixin SSHNPLocalFileMixin on SSHNPImpl {
   late final String sshHomeDirectory;
+
+  final bool _isValidPlatform =
+      Platform.isLinux || Platform.isMacOS || Platform.isWindows;
 
   @override
   Future<void> init() async {
@@ -10,7 +15,12 @@ mixin SSHNPLocalFileMixin on SSHNPImpl {
       throw SSHNPError(
           'The current client type requires allowLocalFileSystem to be true: $runtimeType');
     }
+    if (!_isValidPlatform) {
+      throw SSHNPError(
+          'The current platform is not supported: ${Platform.operatingSystem}');
+    }
     await super.init();
+    
     if (initializedCompleter.isCompleted) return;
   }
 }
