+++ packages/noports_core/lib/src/sshnpd/sshnpd_impl.dart
@@ -8,8 +8,9 @@ import 'package:dartssh2/dartssh2.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:noports_core/src/common/default_args.dart';
+import 'package:noports_core/src/common/file_system_utils.dart';
 import 'package:noports_core/src/common/supported_ssh_clients.dart';
-import 'package:noports_core/src/common/utils.dart';
+import 'package:noports_core/src/common/validation_utils.dart';
 import 'package:noports_core/src/sshrv/sshrv.dart';
 import 'package:noports_core/sshnpd.dart';
 import 'package:noports_core/version.dart';
