+++ packages/noports_core/lib/utils.dart
@@ -1,5 +1,5 @@
 library noports_core_utils;
 
-export 'src/common/utils.dart';
-export 'src/sshnp/utils.dart';
+export 'src/common/validation_utils.dart';
+export 'src/common/file_system_utils.dart';
 export 'src/common/default_args.dart';
